Issue #33, #35: rename temp_config.py -> utils.py to make room for other setup functions; fix output capture and assert no page option when running test_no_profiles()
